[hotfix] serializer 타입 오류 수정

+++ backend/statistics/src/main/java/com/avocado/statistics/kafka/config/ProducerConfig.java
@@ -1,10 +1,8 @@
 package com.avocado.statistics.kafka.config;
 
 import com.avocado.AdStatus;
-import com.avocado.statistics.db.mysql.entity.jpa.Click;
 import io.confluent.kafka.serializers.KafkaAvroSerializer;
-import io.confluent.kafka.streams.serdes.avro.SpecificAvroSerializer;
-import org.apache.kafka.common.serialization.LongSerializer;
+import org.apache.kafka.common.serialization.IntegerSerializer;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -21,13 +19,11 @@ public class ProducerConfig {
     @Value(value = "${spring.kafka.bootstrap-servers}")
     private String bootstrapAddress;
 
-
-
     @Bean
     public KafkaTemplate<Integer, AdStatus> AdStatusKafkaTemplate() {
         Map<String, Object> props = new HashMap<>();
         props.put(org.apache.kafka.clients.producer.ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapAddress);
-        props.put(org.apache.kafka.clients.producer.ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, LongSerializer.class);
+        props.put(org.apache.kafka.clients.producer.ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, IntegerSerializer.class);
         props.put(org.apache.kafka.clients.producer.ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, KafkaAvroSerializer.class);
         props.put("schema.registry.url", "http://a5ef82e13fcbc44689f93c4924981608-494875664.ap-northeast-2.elb.amazonaws.com:8081");
 //        props.put("specific.avro.reader", "true");
@@ -36,7 +32,6 @@ public class ProducerConfig {
 
         return new KafkaTemplate<>(adStatusProducerFactory);
     }
-
